Move `chunk_message` from `utils` to `message`
