+++ framewirc/message.py
@@ -1,5 +1,7 @@
+from collections import deque
+
 from . import commands, exceptions
-from .utils import chunk_message, LINEFEED, to_bytes, to_unicode
+from .utils import LINEFEED, to_bytes, to_unicode
 
 
 ACTION_START = b'\1ACTION '
@@ -69,6 +71,56 @@ def build_message(command, *args, prefix=b'', suffix=b''):
         raise exceptions.MessageTooLong
 
     return message
+
+
+def _chunk_message(message, max_length):
+    # Split the message on linebreaks, and loop over lines.
+    lines = deque(message.splitlines())
+    while lines:
+        line = lines.popleft()
+        line_bytes = to_bytes(line)
+        # If the line fits, add it the the lines.
+        if len(line_bytes) < max_length:
+            yield line_bytes
+            continue
+
+        # Whole line doesn't fit, so see if it can be split on space.
+        letterpoint = None  # Where we should break if there is no space
+        spacepoint = None  # Where we should break if there is a space.
+        line_length = 0  # The running total size of the line
+        for i, str_char in enumerate(line):
+            char_length = len(to_bytes(str_char))
+            line_length += char_length
+            if str_char == ' ':
+                spacepoint = i
+            if line_length > max_length:
+                letterpoint = i
+                break
+
+        if spacepoint is not None:
+            # Break on the last space that fits.
+            start = line[:spacepoint]
+            yield to_bytes(start)
+            # ... and add what's left back into the line pool.
+            end = line[(spacepoint + 1):]
+            lines.appendleft(end)
+            continue
+
+        # Whole line does not contain spaces, so split within word.
+        start = line[:letterpoint]
+        yield to_bytes(start)
+        end = line[letterpoint:]
+        lines.appendleft(end)
+
+
+def chunk_message(message, max_length):
+    """
+    Chunk a unicode message into lines with a max_length in bytes.
+
+    Splits the message by linebreak chars, then words, and finally letters to
+    keep the string chunks short enough.
+    """
+    return list(_chunk_message(message, max_length))
 
 
 def make_privmsgs(target, message, third_person=False):
